Have checkSpaceTest.sh exit as skipped when on Windows
Managing line-ending issues for this one isn't worth it.

+++ scripts/maint/checkSpaceTest.sh
@@ -7,6 +7,19 @@
 umask 077
 set -e
 
+# Skip this test if we're running on Windows; we expect line-ending
+# issues in that case.
+case "$(uname -s)" in
+    CYGWIN*) WINDOWS=1;;
+    MINGW*) WINDOWS=1;;
+    MSYS*) WINDOWS=1;;
+    *) WINDOWS=0;;
+esac
+if test "$WINDOWS" = 1; then
+    # This magic value tells automake that the test has been skipped.
+    exit 77
+fi
+
 # make a safe space for temporary files
 DATA_DIR=$(mktemp -d -t tor_checkspace_tests.XXXXXX)
 trap 'rm -rf "$DATA_DIR"' 0
